requirements.txt
chainlit==0.7.700
cohere==4.37
tiktoken==0.5.2
python-dotenv==1.0.0
langchain==0.1.16
langchain-core ==0.1.46
langchain-community==0.0.34
langchain-openai==0.1.3
qdrant-client==v1.9.0
pymupdf==1.24.2